Multiple bug fixes
  - Print results before flushing streams if server timer procs before
    test end (grace period timeout)
  - Update error logging to allow multiple error string logs in the
    same JSON array
  - Log TCP stream block start and end average counters
  - Fix TCP stream running average calculation

+++ src/iperf_tcp.c
@@ -148,8 +148,10 @@ iperf_tcp_recv(struct iperf_stream *sp)
             int old_stream_counter = sp->result->stream_avg_cntr_blk_strt;
             sp->result->stream_avg_cntr_blk_strt++;
 
+            /// TODO: Add time_diff samples to array in json output
+
             sp->result->stream_avg_tx_to_rx_time_blk_strt =
-                (sp->result->stream_avg_tx_to_rx_time_blk_strt * (old_stream_counter / sp->result->stream_avg_cntr_blk_strt)) +
+                (sp->result->stream_avg_tx_to_rx_time_blk_strt * ( (double)old_stream_counter / (double)sp->result->stream_avg_cntr_blk_strt)) +
                 (time_diff_in_secs / sp->result->stream_avg_cntr_blk_strt);
 
         }
@@ -226,7 +228,7 @@ iperf_tcp_recv(struct iperf_stream *sp)
                 sp->result->stream_avg_cntr_blk_end++;
 
                 sp->result->stream_avg_tx_to_rx_time_blk_end =
-                    (sp->result->stream_avg_tx_to_rx_time_blk_end * (old_stream_counter / sp->result->stream_avg_cntr_blk_end)) +
+                    (sp->result->stream_avg_tx_to_rx_time_blk_end * ((double)old_stream_counter / (double)sp->result->stream_avg_cntr_blk_end)) +
                     (time_diff_in_secs / sp->result->stream_avg_cntr_blk_end);
             }
